### 2019-04-14 #1383 @labarba
@sanskriti-s — Please check the formatting of the metadata section of your paper, and adjust according to the example in the [Author Guide](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography).
