+++ paper/paper.md
@@ -9,13 +9,13 @@ authors:
  - name: Sanskriti Sharma
    orcid: 0000-0002-9884-7351
    affiliation: 1
-- name: Hernan Gelaf-Romer
+ - name: Hernan Gelaf-Romer
    orcid: N/A
    affiliation: 1
-- name: Travis Kessler
+ - name: Travis Kessler
    orcid: 0000-0002-7363-4050
    affiliation: 1
-- name: John Hunter Mack
+ - name: John Hunter Mack
    orcid: N/A
    affiliation: 1
 affiliations:
