### 2019-04-14 #1383 @sanskriti-s
Hi,

I fixed it.

Thank you,
Sanskriti

> On Apr 14, 2019, at 1:40 PM, Lorena A. Barba <notifications@github.com> wrote:
>
> @sanskriti-s <https://github.com/sanskriti-s> — Please check the formatting of the metadata section of your paper, and adjust according to the example in the Author Guide <https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography>.
>
