### 2019-04-14 #1383 @labarba
@sanskriti-s — We do need an ORCID for every author. If your co-author doesn't have one, they will have to create one …
