### 2019-04-15 #1383 @arfon
> @sanskriti-s — We do need an ORCID for every author. If your co-author doesn't have one, they will have to create one …

Actually we don't need ORCIDs for authors, it's just the current value's of `N/A` aren't valid :-) . I've fixed this here: https://github.com/ECRL/ecabc/pull/44
