### 2019-04-22 #1383 @danielskatz
👋 @sanskriti-s - in your code to include the image, you have a leading slash that shouldn't be there.  I think https://github.com/ECRL/ecabc/pull/45 should fix this.  Once you've merged it, or made the same change, you can enter ```@whedon generate pdf``` here in a new comment to try the processing again.
