### 2019-04-30 #1383 @annoviko
Hello, @arocem,I am little bit surprised to be suggested as a reviewer for this submission, because I am not even in the reviewer list. But anyway I have read the paper and have looked through the repository, so if there are no any other candidates, I can review it.1:39, 30 апреля 2019 г., Ariel Rokem <notifications@github.com>:Let me see if we can find someone other than me.
@annoviko: would you be willing to review this JOSS submission?

—You are receiving this because you were mentioned.Reply to this email directly, view it on GitHub, or mute the thread.

Best regards, Andrei
