### 2019-04-30 #1420 @whedon
[REVIEW]: ECabc: A feature tuning program focused on Artificial Neural Network hyperparameters
**Submitting author:** @sanskriti-s (<a href="http://orcid.org/0000-0002-9884-7351">Sanskriti Sharma</a>)
**Repository:** <a href="https://github.com/ECRL/ecabc" target ="_blank">https://github.com/ECRL/ecabc</a>
**Version:** v2.2.2
