### 2019-05-06 #1420 @annoviko
Hello, @sanskriti-s
There are several remarks:
1. Example contains mistakes in README.md (at least it doesn't work for v2.2.2):
- Incorrect import, I think `from ecabc import *` should be instead of `from eabc import *`, otherwise following exception is obtained:
```
ImportError: No module named 'eabc'
```
- ABC is not defined:
```python
    abc = ABC(fitness_fxn=idealDayTest,
            value_ranges=values
            )
```
```
NameError: name 'ABC' is not defined
```

2. Several references does not contains DOI. Could you please add DOI for these references?
```
@online{ecabc_github,
author = {Sanskriti Sharma},
title = {ECabc: A feature tuning program focused on Artificial Neural Network hyperparameters},
year = {2019},
url = {https://github.com/ECRL/ECabc},
urldate = {2019-04-12}
}

@online{ecabc_pypi,
author = {Sanskriti Sharma, Hernan Gelaf-Romer},
title = {ecabc 2.2.2},
year = {2019},
url = {https://pypi.python.org/pypi/ecabc/},
urldate = {2019-04-12}
}

@online{uml_ecrl,
author = {John Hunter Mack},
title = {Energy and Combustion Research Laboratory},
year = {2019},
url = {http://faculty.uml.edu/Hunter_Mack/},
urldate = {2019-04-12}
}

@book{Hassoun_2009,
Author = {{Hassoun}, M.},
Booktitle = {Fundamentals of Artificial Neural Networks, by Hassoun, Mohamad H.~ISBN 9788120313569 (HB).~Published by PHI Learning, Delhi, India, 2009.},
Publisher = {PHI Learning},
Title = {{Fundamentals of Artificial Neural Networks}},
Year = 2009
}
```
