### 2019-05-15 232555c
fixed README typos
