+++ README.md
@@ -46,7 +46,7 @@ Additional package dependencies (Numpy) will be installed during the ECabc insta
 
 To get started import ECabc
 ```python
-from ecabc import *
+from ecabc.abc import *
 ```
 Then define your fitness function as a function. The fitness function is the user defined function whose solution is being optimized. Pass in the values and args and have it return the output that is being optimized
 ```python
@@ -137,7 +137,7 @@ The goal is to have the first two values added up to as close to 70 as possible,
 close to 37.5 as possible.
 '''
 
-from eabc import *
+from ecabc.abc import *
 import os
 import time
 
