### 2019-05-15 #1420 @sanskriti-s
Hi @annoviko. Sorry for the late response, I had finals. So I fixed the README, but I have a question about the DOIs. To the best of my knowledge online sources don't have DOIs and while books do, I can't find one for the book I have cited. So what should I do about them?
