### 2019-05-22 #1420 @arokem
Hi @sanskriti-s : you are correct that the websites you refer do not have a DOI, and you do not need to specify one for these. I believe that the Hassoun book does not have a DOI either. So, no need for DOIs for these.
