### 2019-06-01 #1420 @annoviko
Hello, @sanskriti-s ,
There is only one unclear moment that is related to testing. I haven't found any tests in the repository and there is no instruction how to verify software. Could you please provide any comments?
