### 2019-06-04 #1420 @annoviko
Hello, @sanskriti-s ,

I think it should be reflected in the README in line with JOSS's requirements:
> Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified?
Could you please add this information (how to verify software functionality) to README?
