### 2019-06-05 e2820f4
Add testing to README
