### 2019-06-13 #1420 @arokem
@sanskriti-s : just one request on my end: the caption of Figure one is currently not particularly informative. Would you mind rewriting this caption? Even just "Figure 1: the stages of the algorithm (see text)" or similar would be sufficient.

After that's done, we can proceed with the next steps in accepting the article.
