### 2019-06-20 4b068be
Added caption to figure 1
