### 2019-06-24 c50271e
Associate the caption with the figure.
