### 2019-06-27 #1420 @arokem
Great. I think this is ready for your approval @openjournals/joss-eics
