### 2019-06-27 #1420 @kyleniemeyer
Hi @sanskriti-s, please clean up the Zenodo archive's metadata—the author list should match that of the paper, and right now it includes "KOLANICH; Daniel S. Katz; Ariel Rokem; Arfon Smith" and is missing Hunter Mack.
