### 2019-06-28 #1420 @kyleniemeyer
@sanskriti-s there are a few other changes needed in the paper:

- Please add more-complete affiliation information, matching what you'd see in any journal article (department, institution, city, state, country)
- In the first paragraph, please fix the citations using the appropriate commands:
```
- `@author:2001`  ->  "Author et al. (2001)"
- `[@author:2001]` -> "(Author et al., 2001)"
- `[@author1:2001; @author2:2001]` -> "(Author1 et al., 2001; Author2 et al., 2002)"
```
In particular, I would change `research by Z. Cam et al [@Cam_2015]` to `research by @Cam_2015`.
- In the third paragraph, please add a comma after `i.e.`
- Please revise the last sentence of the third paragraph, it seems incomplete. `Alternatively, a number of iterations can be and for each iteration...`, "a number of iterations" can be what?
- Please fix the `Cam_2015` reference info, it looks malformed. The `author` field should include all of the author names.
- Actually, it looks like many of the references only include the first author and "et al.". Please add the full author lists for each citation, and remove all instances of "et al."
- The Mack reference is not appropriate (you're citing a research group page); either remove the citation and just mention the group in the text, or add citations to actual publications that have used the software.
