### 2019-07-11 8183502
changes to paper
