+++ paper/paper.bib
@@ -3,7 +3,11 @@ author = {Sanskriti Sharma},
 title = {ECabc: A feature tuning program focused on Artificial Neural Network hyperparameters},
 year = {2019},
 url = {https://github.com/ECRL/ECabc},
-urldate = {2019-04-12}
+urldate = {2019-04-12},
+institution = {University of Massachuetts Lowell},  
+city = {Lowell},
+state = {Massachusetts},
+country = {United States of America}
 }
 
 @online{ecabc_pypi,
@@ -11,58 +15,71 @@ author = {Sanskriti Sharma, Hernan Gelaf-Romer},
 title = {ecabc 2.2.2},
 year = {2019},
 url = {https://pypi.python.org/pypi/ecabc/},
-urldate = {2019-04-12}
-}
-
-@online{uml_ecrl,
-author = {John Hunter Mack},
-title = {Energy and Combustion Research Laboratory},
-year = {2019},
-url = {http://faculty.uml.edu/Hunter_Mack/},
-urldate = {2019-04-12}
+urldate = {2019-04-12},
+institution = {University of Massachuetts Lowell},  
+city = {Lowell},
+state = {Massachusetts},
+country = {United States of America}
 }
 
 @book{Hassoun_2009,
-Author = {{Hassoun}, M.},
-Booktitle = {Fundamentals of Artificial Neural Networks, by Hassoun, Mohamad H.~ISBN 9788120313569 (HB).~Published by PHI Learning, Delhi, India, 2009.},
-Publisher = {PHI Learning},
-Title = {{Fundamentals of Artificial Neural Networks}},
-Year = 2009
+author = {{Hassoun}, M.},
+booktitle = {Fundamentals of Artificial Neural Networks, by Hassoun, Mohamad H.~ISBN 9788120313569 (HB).~Published by PHI Learning, Delhi, India, 2009.},
+publisher = {PHI Learning},
+title = {{Fundamentals of Artificial Neural Networks}},
+year = 2009
 }
 
 @article{Castillo_2000,
-author = {P. Castillo, et al},
+author = {P. Castillo, J. Merelo, A. Prieto, V. Rivas, G. Romero},
 title = "{G-Prop: Global Optimization of Multilayer Perceptrons Using GAs}",
-journal = {Neurocomputer},
-year = 2000,
-month = may,
-volume = 35,
+journal = {Neurocomputing},
+year = {2000},
+month = {november},
+volume = {35},
 doi = {10.1016/s0925-2312(00)00302-7},
+department = {Departamento de Arquitectura y Tecnologı́a de Computadoras},
+institution = {Universidad de Granada},  
+city = {Granada},
+country = {Spain}
 }
 
 @proceedings{Cam_2015,
-author = {Z. Cam et al},
+author = {Z. Cam, S. Yetis, T. Yildirim},
 title = {Learning Parameter Optimization of Multi-Layer Perceptron Using Artificial Bee Colony, Genetic Algorithm and Particle Swarm Optimization},
 booktitle = {2015 IEEE 13th International Symposium on Applied Machine Intelligence and Informatics},
 publisher = {IEEE},
 year = {2015},
-doi = {10.1109/sami.2015.7061899}
+month = {march},
+doi = {10.1109/sami.2015.7061899},
+isbn = {978-1-4799-8221-9},
+institution = {Yildiz Technical University/Electronics & Communications Engineering},
+city = {Istanbul},
+country = {Turkey}
 }
 
 @article{Kessler_2017_1,
-author = {T. Kessler, et al},
+author = {T. Kessler, J. Mack},
 title = {ECNet: Large scale machine learning projects for fuel property prediction},
 journal = {Journal of Open Source Software},
 year = {2017},
[…]
