+++ paper/paper.md
@@ -10,11 +10,13 @@ authors:
    orcid: 0000-0002-9884-7351
    affiliation: 1
  - name: Hernan Gelaf-Romer
+   orcid: 0000-0003-0690-576X
    affiliation: 1
  - name: Travis Kessler
    orcid: 0000-0002-7363-4050
    affiliation: 1
  - name: John Hunter Mack
+   orcid: 0000-0002-5455-8611
    affiliation: 1
 affiliations:
  - name: UMass Lowell Energy and Combustion Research Laboratory
@@ -22,18 +24,18 @@ affiliations:
 date: 12 April 2019
 bibliography: paper.bib
 nocite: |
- @ecabc_github, @ecabc_pypi, @uml_ecrl
+ @ecabc_github, @ecabc_pypi
 ---
 
 # Summary
 
-A rich body of literature exists regarding the optimization of artificial neural networks (ANN), a method comprised of densely interconnected adaptive processing units [@Hassoun_2009]. This need for optimization arises from a large number of user-set parameters that greatly affect the quality of the ANN’s training. Traditionally, this has been done manually. However, presently with higher computing capacity, it is possible to utilize algorithmic approaches which greatly increase the speed of optimization while enhancing model performance. Commonly used optimization routines include genetic algorithms [@Castillo_2000] and particle swarm optimizations [@Cam_2015]. However, research by Z. Cam et al [@Cam_2015] has shown that an Artificial Bee Colony (ABC) performs best by providing higher accuracy and a shorter run time.
+A rich body of literature exists regarding the optimization of artificial neural networks (ANN), a method comprised of densely interconnected adaptive processing units (Hassoun et. al., 2009). This need for optimization arises from a large number of user-set parameters that greatly affect the quality of the ANN’s training. Traditionally, this has been done manually. However, presently with higher computing capacity, it is possible to utilize algorithmic approaches which greatly increase the speed of optimization while enhancing model performance. Commonly used optimization routines include genetic algorithms (Castillo et. al., 2000) and particle swarm optimizations (Cam et. al., 2015). However, research by Z. Cam et. al. (2015) has shown that an Artificial Bee Colony (ABC) performs best by providing higher accuracy and a shorter run time.
 
 ECabc is an open source Python package based on an Artificial Bee Colony, used for tuning functions such as hyperparameters of artificial neural networks. The method is modelled after the honey foraging techniques of bees, where the search space for  solutions is in as many dimensions as the number of parameters being tuned. The algorithm is outlined in Figure 1.
 
 ![Stages of the algorithm](images/state.png)
 
[…]
