### 2019-07-11 #1420 @sanskriti-s
@kyleniemeyer : Fixed paper and bibliography
