### 2019-07-11 #1420 @kyleniemeyer
@sanskriti-s I fixed a number of the remaining issues with the paper and references in https://github.com/ECRL/ecabc/pull/47 and https://github.com/ECRL/ecabc/pull/48

When I was asking for affiliation information, I meant your affiliation given in the paper header, not to add institution to each of the references—which is not necessary (or even used).

Also, it looked like you actually removed the citation commands, rather than correcting them as I asked; I fixed those.

Finally, in BibTeX references, author names in `author` need to be separated by `and` rather than commas.
