### 2019-07-11 ddd4d1f
Update paper.md
Removed extraneous in-text reference that I missed before
