+++ paper/paper.md
@@ -27,7 +27,7 @@ bibliography: paper.bib
 
 # Summary
 
-A rich body of literature exists regarding the optimization of artificial neural networks (ANN), a method comprised of densely interconnected adaptive processing units [@Hassoun_2009]. This need for optimization arises from a large number of user-set parameters that greatly affect the quality of the ANN's training. Traditionally, this has been done manually. However, presently with higher computing capacity, it is possible to utilize algorithmic approaches which greatly increase the speed of optimization while enhancing model performance. Commonly used optimization routines include genetic algorithms [@Castillo_2000] and particle swarm optimizations [@Cam_2015] (Cam et. al., 2015). However, @Cam_2015 showed that an Artificial Bee Colony (ABC) performs best by providing higher accuracy and a shorter run time.
+A rich body of literature exists regarding the optimization of artificial neural networks (ANN), a method comprised of densely interconnected adaptive processing units [@Hassoun_2009]. This need for optimization arises from a large number of user-set parameters that greatly affect the quality of the ANN's training. Traditionally, this has been done manually. However, presently with higher computing capacity, it is possible to utilize algorithmic approaches which greatly increase the speed of optimization while enhancing model performance. Commonly used optimization routines include genetic algorithms [@Castillo_2000] and particle swarm optimizations [@Cam_2015]. However, @Cam_2015 showed that an Artificial Bee Colony (ABC) performs best by providing higher accuracy and a shorter run time.
 
 ECabc is an open source Python package based on an Artificial Bee Colony, used for tuning functions such as hyperparameters of artificial neural networks. The method is modelled after the honey foraging techniques of bees, where the search space for  solutions is in as many dimensions as the number of parameters being tuned. The algorithm is outlined in Figure 1.
 
