### 2019-07-11 #1420 @kyleniemeyer
@sanskriti-s congrats on your submission's publication in JOSS! Thanks to @annoviko for reviewing and @arokem for editing!
